backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-cors-headers==3.7.0
django-extensions==3.1.5
django-filter==21.1
django-seed==0.3.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djhtml==1.4.11
dr-scaffold==2.1.1
drf-yasg==1.20.0
Faker==11.3.0
idna==3.3
inflect==5.3.0
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
Pillow==9.0.0
psycopg2-binary==2.9.3
PyJWT==1.7.1
pyparsing==3.0.6
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toposort==1.7
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
